kv_e8: store secondary indices as uint8 to realise 3/5/7 bpw savings
The 256-entry small codebook only needs 8 bits per index, but the
quantize path was casting the secondary indices to int16 before
returning them — so 3/5/7 bpw recipes ended up taking *the same*
storage as 4/6/(non-existent) bpw. The kv_compress_memory bench
caught this: 3 bpw landed at 22.4 KB/tok, exactly matching 4 bpw,
which made the half-step rungs pointless from a memory standpoint.

Fix: cast secondary indices to torch.uint8 (range 0..255 fits in 8
bits with no sign-extension concern). The downstream ``.long()``
call in dequantize() still produces the correct positive indices.

Re-bench on Gemma-4-E4B-it, ctx=16384:
  3 bpw:  18.8 KB/tok (was 22.4, -16%)
  5 bpw:  26.0 KB/tok (new)
  7 bpw:  33.2 KB/tok (new)
Storage now scales linearly with bpw, +3.5 KB/tok per step.

+++ glq/kv_e8.py
@@ -156,11 +156,15 @@ class E8KVQuantizer:
         # unit-ball inputs because it's just the lowest-norm shells of
         # the primary. Between consecutive secondary stages we re-scale
         # again so each stage sees a residual at the same magnitude.
+        # The small codebook has 256 entries → indices fit in uint8.
+        # Storing them as uint8 (not int16) is what actually realises
+        # the 3-bpw / 5-bpw / 7-bpw storage savings on the cache; the
+        # math is otherwise identical to the int16 layout.
         if self.secondary_stages > 0:
             cur = residual * rs
             for k in range(1, self.secondary_stages + 1):
                 dec_s, idx_s = self.secondary_codebook.quantize(cur)
-                out[f"idx_s{k}"] = idx_s.reshape(N, n_groups).to(torch.int16)
+                out[f"idx_s{k}"] = idx_s.reshape(N, n_groups).to(torch.uint8)
                 if k < self.secondary_stages:
                     cur = (cur - dec_s) * rs
         # Reshape scale back to original leading shape (drop the last
